+++ Changelog.md
@@ -8,6 +8,8 @@
 
 ### 🐛 Bug fixes
 
+- Fix bugs when assigning sections to starter file groups across multiple assignments (#7523)
+
 ### 🔧 Internal changes
 
 - Remove `activerecord-session_store` gem (#7517)